updates to Transaction domain, Transaction repository, Transaction service, Transaction controller and Transaction template

+++ src/main/java/com/codercampus/Assignment11/service/TransactionService.java
@@ -21,7 +21,7 @@ public class TransactionService {
         return transactionRepo.save(transaction);
     }
 
-    public Transaction findById(Integer transactionId) {
+    public Transaction findById(Long transactionId) {
         return transactionRepo.findById(transactionId);
     }
